requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core==1.29.4
azure-functions==1.17.0
azure-identity==1.14.0
azure-storage-blob==12.18.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
mailerlite==0.1.3
MarkupSafe==2.1.3
msal==1.24.0
msal-extensions==1.0.0
portalocker==2.7.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
